chore: move dashboard controller to user. fix statistics route name
