+++ app/Http/Controllers/Api/V1/User/DashboardController.php
@@ -1,9 +1,8 @@
 <?php
 
-namespace App\Http\Controllers\Api\V1\Admin;
+namespace App\Http\Controllers\Api\V1\User;
 
 use App\Http\Controllers\Controller;
-use App\Models\Order;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
@@ -84,14 +83,25 @@ class DashboardController extends Controller
             'message' => 'Dashboard retrieved successfully',
             'status_code' => Response::HTTP_OK,
             'data' => [
-                'current_month_revenue' => $currentMonthRevenue,
-                'difference_in_revenue_percent' => $percentageDifference . '%',
-                'subscriptions' => 0,
-                'monthly_subscriptions_diff' => 0,
-                'current_month_orders' => $currentMonthOrders,
-                'difference_in_orders_percent' => $percentageDifferenceOrders . '%',
-                'active_users_count' => $activeUser,
-                'difference_an_hour_ago' => max(($activeUser - $activeUserAnHourAgo), 0),
+                'revenue' => [
+                    'current_month' => $currentMonthRevenue,
+                    'previous_month' => $lastMonthRevenue, // Assuming you have this data
+                    'percentage_difference' => $percentageDifference . '%',
+                ],
+                'subscriptions' => [
+                    'current_month' => 0,
+                    'previous_month' => 0, // Assuming you have this data
+                    'percentage_difference' => 0 . '%',
+                ],
+                'orders' => [
+                    'current_month' => $currentMonthOrders,
+                    'previous_month' => $lastMonthOrders, // Assuming you have this data
+                    'percentage_difference' => $percentageDifferenceOrders . '%',
+                ],
+                'active_users' => [
+                    'current' => $activeUser,
+                    'difference_an_hour_ago' => max(($activeUser - $activeUserAnHourAgo), 0),
+                ],
             ]
         ]);
     }
